pyproject.toml
[project]
name = "demuxai"
version = "0.1.0"
description = "A local proxy that provides a singular OpenAI compatible API for all your AI services. Try it out with your Jetbrains IDE"
readme = "README.md"
authors = [
    { name = "Blaine Jester", email = "bjester@quavalent.com" }
]
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0.3",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "fire>=0.7.1",
    "uvicorn",
    "dotenv",
]

[project.scripts]
demuxai = "demuxai.cli:main"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pre-commit",
]

[build-system]
requires = ["uv_build>=0.9.17,<0.10.0"]
build-backend = "uv_build"

[tool.pytest.ini_options]
pythonpath = "src"
testpaths = ["tests"]

[tool.flake8]
max-line-length = 100
max-complexity = 20

[tool.uv.build-backend]
source-exclude = [
    ".envrc",
    ".idea/",
    ".git/",
    "dist/"
]
wheel-exclude = [
    ".github/",
    ".gitignore/",
    ".precommit-config.yaml",
    ".venv/",
    "Makefile",
    "tests/"
]
